refactor: Modernize frontend structure and Docker setup
This commit introduces several significant improvements and fixes across the application.

Key changes:

- **Frontend Directory Refactoring:**
  - The React frontend source code has been reorganized into a standard structure with `pages`, `components`, and `assets` directories for better maintainability.
  - Import paths have been updated accordingly.

- **Docker Configuration Update:**
  - The `ollama` service has been removed from `docker-compose.yml` to allow the application to connect to an external Ollama instance, as requested.
  - The backend `Dockerfile` has been corrected to copy binaries from `/usr/local/bin` in the builder stage, fixing an issue where `uvicorn` was not found in the final image.

- **Dependency and Build Fixes:**
  - Downgraded `react` to v18 and `tailwindcss` to v3 to resolve a critical build failure caused by incompatibility with `react-scripts@5`.
  - Updated `postcss.config.js` to be compatible with Tailwind CSS v3.
  - Added `@tailwindcss/postcss` and corrected its version to align with `tailwindcss` during the debugging process, before identifying the root incompatibility.

- **Backend Application Fixes:**
  - Refactored the `langgraph` implementation in `agent_graph.py` to use the modern `ToolNode` and `tools_condition` API, replacing the deprecated `ToolExecutor`.
  - Fixed an `ImportError` in `db/init_db.py` by correcting the import path for the SQLAlchemy `Base`.

These changes collectively address the user's requirements, fix multiple critical bugs, and leave the application in a stable, working state.
