+++ backend/app/services/agent_graph.py
@@ -4,7 +4,7 @@ from typing import TypedDict, Annotated, Sequence
 from langchain_core.messages import BaseMessage, HumanMessage, AIMessage, ToolMessage
 from langchain_openai import ChatOpenAI
 from langgraph.graph import StateGraph, END
-from langgraph.prebuilt import ToolExecutor
+from langgraph.prebuilt import ToolNode, tools_condition
 
 from .agent_tools import all_tools
 
@@ -29,8 +29,8 @@ class AgentState(TypedDict):
 
 # --- 2. Define the Graph Nodes ---
 
-# Initialize the tool executor and the language model
-tool_executor = ToolExecutor(all_tools)
+# Initialize the ToolNode with our set of tools
+tool_node = ToolNode(all_tools)
 
 model = ChatOpenAI(
     temperature=0,
@@ -51,49 +51,7 @@ def call_model(state: AgentState):
     # The response from the model is added to the state
     return {"messages": [response]}
 
-def call_tool(state: AgentState):
-    """
-    The 'action' node. It executes the tools that the model has decided to call.
-    """
-    print("--- ACTION: Calling tool... ---")
-    # The last message in the state should be the AI's decision to call a tool
-    last_message = state['messages'][-1]
-
-    # Ensure the message is an AIMessage with tool_calls
-    if not isinstance(last_message, AIMessage) or not last_message.tool_calls:
-        # This case should ideally not be reached if the graph logic is correct
-        return {"messages": [HumanMessage(content="Error: No tool call found in the last message.")]}
-
-    # Execute the tool calls
-    tool_invocations = last_message.tool_calls
-    tool_outputs = tool_executor.batch(tool_invocations)
-
-    # Format the outputs as ToolMessage objects
-    tool_messages = [
-        ToolMessage(content=str(output), tool_call_id=inv["id"])
-        for inv, output in zip(tool_invocations, tool_outputs)
-    ]
-
-    return {"messages": tool_messages}
-
 # --- 3. Define the Graph Edges ---
-
-def should_continue(state: AgentState) -> str:
-    """
-    This function decides the next step after the model has been called.
-    - If the model decided to call a tool, we route to the 'action' node.
-    - If the model responded directly without calling a tool, we are done ('END').
-    """
-    print("--- DECISION: Evaluating model response... ---")
-    last_message = state['messages'][-1]
-    if isinstance(last_message, AIMessage) and last_message.tool_calls:
-        # The model wants to use a tool
-        print("--- DECISION: Route to ACTION ---")
-        return "continue"
-    else:
-        # The model has provided a final answer
-        print("--- DECISION: Route to END ---")
-        return "end"
 
 # --- 4. Assemble the Graph ---
 
@@ -102,19 +60,25 @@ workflow = StateGraph(AgentState)
 
 # Add the nodes
 workflow.add_node("agent", call_model)
-workflow.add_node("action", call_tool)
+workflow.add_node("tools", tool_node)
 
 # Define the entry point and the edges
 workflow.set_entry_point("agent")
+
+# Use the prebuilt tools_condition to decide whether to call tools or end
 workflow.add_conditional_edges(
     "agent",
-    should_continue,
+    tools_condition,
     {
-        "continue": "action",
+        # If the agent wants to call a tool, route to the 'tools' node
+        "continue": "tools",
+        # If the agent is finished, route to the 'end' node
         "end": END,
     },
 )
-workflow.add_edge("action", "agent")
+
+# After calling a tool, we route back to the agent to respond
+workflow.add_edge("tools", "agent")
 
 # Compile the graph into a runnable object
 agent_graph = workflow.compile()
