+++ docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.8'
-
 services:
   backend:
     build:
@@ -20,8 +18,6 @@ services:
       - .env
     ports:
       - "8000:8000"
-    depends_on:
-      - ollama
     networks:
       - ecommerce-net
 
@@ -36,17 +32,6 @@ services:
       - "3000:80" # Map host port 3000 to container port 80 (Nginx)
     depends_on:
       - backend
-    networks:
-      - ecommerce-net
-
-  ollama:
-    image: ollama/ollama
-    container_name: ollama
-    restart: unless-stopped
-    volumes:
-      - ollama_data:/root/.ollama
-    ports:
-      - "11434:11434"
     networks:
       - ecommerce-net
